throw when upload file not support
